requirements.txt
accelerate==0.32.1
einops==0.8.0
gradio==4.38.1
joblib==1.2.0
numpy==1.26.4
omegaconf==2.3.0
packaging==24.1
pandas==1.5.3
peft==0.11.1
Pillow==9.4.0
pycocoevalcap==1.2
pycocotools==2.0.8
pynvml==11.5.0
Requests==2.32.3
scikit_learn==1.5.1
spacy==3.7.5
timm==0.4.12
torch==2.3.1
torchmetrics==0.11.4
torchvision==0.18.1
tqdm==4.64.1
transformers==4.42.4
webdataset==0.2.86
fairscale==0.4.4